requirements.txt
alabaster==0.7.12
asgiref==3.5.2
attrs==22.1.0
Babel==2.10.3
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-allauth==0.51.0
django-bootstrap4==22.2
django-crispy-forms==1.14.0
docutils==0.17.1
gunicorn==20.1.0
idna==3.3
imagesize==1.4.1
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psutil==5.9.1
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-services==2.2.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
zc.lockfile==2.0
